requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
filelock==3.0.12
flasgger==0.9.5
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
numpy==1.19.5
oauthlib==3.1.1
openapi-codec==1.3.2
opt-einsum==3.3.0
platformdirs==2.1.0
protobuf==3.18.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.18.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.6.0
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
